ai_matching/requirements.txt
fastapi
uvicorn
numpy
pydantic
sentence-transformers
scikit-learn
torch
rapidfuzz
pdfplumber
spacy
faiss-cpu
python-multipart
